remove bind_kv_cache
